Fixed minor type issue on Windows

+++ src/BitBoard.h
@@ -336,9 +336,10 @@ public:
 	inline u32 bitScanForward(const u64 piece)
 	{
 		#if defined(_WIN32)
-		u32 index;
+		unsigned long int index;
 		_BitScanForward64(&index, piece);
 		return index;
+
 		#elif defined(__gnu_linux__) || defined(__linux__) || defined(__CYGWIN__)
 		return __builtin_ctzll(piece);
 		#endif
